+++ packages/mobile/src/chips/__figma__/TabbedChips.figma.tsx
@@ -10,7 +10,7 @@ figma.connect(
   {
     imports: ["import { TabbedChips } from '@coinbase/cds-mobile/chips/TabbedChips';"],
     props: {
-      overflow15050: figma.boolean('overflow'),
+      overflow: figma.boolean('overflowing'),
       platform: figma.enum('platform', {
         mobile: 'mobile',
         desktop: 'desktop',